+++ lib/whois/record/parser/whois.dreamhost.com.rb
@@ -19,7 +19,7 @@ module Whois
       # @see Whois::Record::Parser::Example
       #   The Example parser for the list of all available methods.
       #
-      # @since NEXT_RELEASE
+      # @since 2.4.0
       class WhoisDreamhostCom < Base
 
         property_not_supported :status